generate codegen lab
