+++ codegen/codegen_lab.ts
@@ -1,0 +1,6 @@
+// Generated lab file. DO NOT MODIFY.
+import { Lab } from "labs/labs.ts";
+import { generateLab } from "labs/codegen/generate.ts"
+
+export const codegenLab = new Lab()
+  .procedure("codegen.lab", generateLab);
