+++ deno.json
@@ -3,7 +3,7 @@
     "labs/": "./"
   },
   "tasks": {
-    "example": "deno run ./example/example.ts",
-    "generate:example": "deno run -Ar https://deno.land/x/generate/cli/main.ts ./example/gen.ts"
+    "generate": "deno run -Ar https://deno.land/x/generate/cli/main.ts gen.ts",
+    "example": "deno run ./example/example.ts"
   }
 }
