--- example/gen.ts
@@ -1,1 +1,0 @@
-//deno:generate deno run --allow-write codegen.ts
