+++ example/my_lab.ts
@@ -1,3 +1,4 @@
+// Generated lab file. DO NOT MODIFY.
 import { Lab } from "labs/labs.ts";
 import { linksLab } from "labs/links.ts"
 import { notesLab } from "labs/notes.ts"
