+++ gen.ts
@@ -1,0 +1,2 @@
+//deno:generate deno run --allow-write codegen/codegen.ts
+//deno:generate deno run --allow-write example/codegen.ts
